requirements.txt
numpy==1.21.5
opencv-python==4.5.5.64
Pillow==9.1.0
streamlit==1.8.1
